requirements.txt
pydub==0.23.1
PyAudio~=0.2.13
numpy~=1.24.3
scipy~=1.10.1
matplotlib~=3.7.1
mysql-connector-python==8.0.17
psycopg2==2.8.3

Django~=4.2.2
djangorestframework~=3.14.0
pyacoustid~=1.2.2
acrcloud~=2.0